organizes sectors module
