+++ src/services/__init__.py
@@ -1,9 +1,9 @@
 from src.services.quotes.get_quotes import get_quotes, get_simple_quotes
+from src.services.sectors.get_sectors import get_sectors, get_sector_for_symbol, get_sector_details
 from .get_historical import get_historical
 from .get_indices import scrape_indices
 from .get_movers import scrape_actives, scrape_gainers, scrape_losers
 from .get_news import scrape_news_for_quote, scrape_general_news
 from .get_search import get_search
-from .get_sectors import get_sectors, get_sector_for_symbol, get_sector_details
 from .get_similar_quotes import scrape_similar_quotes
 from .indicators import get_summary_analysis
