+++ src/services/sectors/__init__.py
@@ -1,0 +1,2 @@
+from .fetchers.sector_scraper import parse_sector, parse_sector_details
+from .utils import get_yahoo_sector
